+++ include/RaylibFrontend.h
@@ -21,8 +21,8 @@ private:
     Raytracer& raytracer;
     Texture2D textureOnGpu;
     std::chrono::time_point<std::chrono::steady_clock> now;
-    double cameraMovSpeed{10.0};
-    double cameraRotSpeed{30.0};
+    float cameraMovSpeed{10.0};
+    float cameraRotSpeed{30.0};
     bool notifyAboutFrameCompletion{true};
 
     void input();
